services/users/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
PyJWT==2.8.0
gunicorn==21.2.0
alembic==1.12.0
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
python-magic==0.4.27
moviepy==1.0.3

